@@ -1,10 +1,25 @@
+// Copyright (C) 2022 Jared Allard <jared@rgst.io>
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 package config
 
 import (
-	"io/ioutil"
+	"os"
 
 	"github.com/pkg/errors"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 var (
@@ -16,6 +31,9 @@ type Cloud string
 
 // ProxyConfig is a configuration file for the proxy
 type ProxyConfig struct {
+	// ListenAddress is the address to listen on (the proxy)
+	ListenAddress string `yaml:"listenAddress"`
+
 	// The Cloud this instance is in
 	Cloud       Cloud         `yaml:"cloud"`
 	Server      *ServerConfig `yaml:"server"`
@@ -58,7 +76,7 @@ type DockerConfig struct {
 
 // LoadProxyConfig loads a proxy configuration file
 func LoadProxyConfig(path string) (*ProxyConfig, error) {
-	b, err := ioutil.ReadFile(path)
+	b, err := os.ReadFile(path)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to read config file")
 	}
