@@ -1,4 +1,19 @@
-package instance
+// Copyright (C) 2022 Jared Allard <jared@rgst.io>
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+package gcp
 
 import (
 	"context"
@@ -8,6 +23,7 @@ import (
 	"github.com/jaredallard/minecraft-preempt/pkg/cloud"
 	"golang.org/x/oauth2/google"
 	"google.golang.org/api/compute/v1"
+	"google.golang.org/api/option"
 )
 
 var (
@@ -33,7 +49,7 @@ func NewClient(ctx context.Context, project, zone string) (*Client, error) {
 		return nil, err
 	}
 
-	comp, err := compute.New(client)
+	comp, err := compute.NewService(ctx, option.WithHTTPClient(client))
 	if err != nil {
 		return nil, err
 	}
